Start of a refactoring project to expose more semantic info about property values
Property values are exposed as strings via the Value property but this leaves out a lot of the semantic information that was gleaned during the parsing process - for example, if a colour value was parsed from a hex string or a colour function, that information is not easily retrieved since the colours are canonicalised to rgb(....) and even then, there is no easy way to get to the Color instance that was constructed internally to get things like the R/G/B values etc.

This attempts to improve on that by introducing a new IValue abstraction that can expose various data types that have been built during the parsing and value conversion process. Each IValue has a "Kind" that indicates what type of value it represents and based on that, the actual value object itself can be extracted via .As<T> from that IValue.

Another change is to make the value conversion process more flexible so that everything doesn't have to be resolved via a IValueConverter. This helps in those cases where the value of a property is ambiguous - for example consider the "flex" property shorthand where in its single value form can accept a value for flex-basis or flex-grow depending on the actual value. A new CoerceValue method has been added to Property that can be overriden in those properties which may require more complex control flow/logic to deal with value conversion in cases like this.

Finally, the existing Value property (which was a string), has been renamed to ValueText to better reflect what it is.

This initial change has converted the BackgroundColour property over to the new infrastructure which is contained with the \New folder for now as additional properties are migrated over.