Refactoring of the new property value types
The idea here is to have the value classes be self-contained and inherited from BaseValue which is itself inherited from IValue. This requires changing some of the existing value classes such as Color etc. to not be a struct and to accept and additional constructor parameter for the parsed value which had a knock-on effect to a bunch oif places